pyproject.toml
[tool.poetry]
name = "video_streamming"
version = "0.1.0"
description = ""
authors = ["xal9wiii4ik <xal9wiii4ik@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.26.0"
Flask = "^2.0.2"
python-dotenv = "^0.19.2"
Flask-SQLAlchemy = "^2.5.1"
psycopg2-binary = "^2.9.2"
gunicorn = "^20.1.0"
pre-commit = "^2.16.0"
Flask-Pydantic = "^0.9.0"
PyJWT = "^2.3.0"
Flask-Cors = "^3.0.10"
Flask-Migrate = "^3.1.0"

[tool.poetry.dev-dependencies]
pre-commit = "^2.16.0"
mypy = "^0.930"
flake8 = "^4.0.1"
ipython = "^7.30.1"
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
